Commit producer (+update) spring boot 3.3.3 swagger ui Debug

+++ src/main/java/org/example/event/ProductEvent.java
@@ -23,7 +23,7 @@ public class ProductEvent {
     private String business_address;
     private String address;
     private Integer rcbic;
-    private Float corr_ass;
-    private Float ass;
+    private Double corr_ass;
+    private Double ass;
     private String bank_name;
 }
